refactor(utils): move getCategoryChartData from lib/chart to lib/utils/chart
